feat: add cli
