+++ tsconfig.json
@@ -6,6 +6,7 @@
     "outDir": "dist",
     "resolveJsonModule": true,
     "noUnusedLocals": false,
+    "skipLibCheck": true,
     "lib": ["DOM"]
   },
   "include": ["./src/**/*", "./config.ts"]
